Update rating, add paging for comments

+++ MyShop/Models/Product.cs
@@ -23,7 +23,7 @@ namespace MyShop.Models
         public string MetaDiscription { get; set; } = String.Empty;
 
         //[Column(TypeName = "decimal(5,1)")]
-        public double? Rating { get; set; } = 5;
+        public float? Rating { get; set; } = 5f;
         public bool Visible { get; set; } = true;
         public bool Deleted { get; set; } = false;
         [NotMapped]
